+++ lib/crawler/options.ex
@@ -10,6 +10,7 @@ defmodule Crawler.Options do
   @workers 10
   @interval 0
   @max_depths 3
+  @max_pages :infinity
   @timeout 5_000
   @store nil
   @user_agent "Crawler/#{Mixfile.project()[:version]} (https://github.com/fredwu/crawler)"
@@ -44,6 +45,7 @@ defmodule Crawler.Options do
         workers: workers(),
         interval: interval(),
         max_depths: max_depths(),
+        max_pages: max_pages(),
         timeout: timeout(),
         store: store(),
         user_agent: user_agent(),
@@ -85,6 +87,7 @@ defmodule Crawler.Options do
   defp workers, do: Application.get_env(:crawler, :workers, @workers)
   defp interval, do: Application.get_env(:crawler, :interval, @interval)
   defp max_depths, do: Application.get_env(:crawler, :max_depths, @max_depths)
+  defp max_pages, do: Application.get_env(:crawler, :max_pages, @max_pages)
   defp timeout, do: Application.get_env(:crawler, :timeout, @timeout)
   defp store, do: Application.get_env(:crawler, :store, @store)
   defp user_agent, do: Application.get_env(:crawler, :user_agent, @user_agent)