+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## master
 
 - [Added] `:store` option, defaults to `nil` to save memory usage
+- [Added] `:max_pages` option
 
 ## v1.2.0 [2023-09-29]